requirements.txt
streamlit
plotly
pandas
numpy
scikit-learn
torch
matplotlib
kaleido
nbformat
openai
groq
google-generativeai
ollama
